UFF - Ultrasound File Format

Points

USTB.UFF.Point — Type

Point(r, θ, ϕ)

Point contains the position of a point in a tridimensional space. It express that location in spherical coordinates which allows to place points at infinity but in a given direction.

The Julia implementation of the UFF Point type is a derivation of the CoordinateTransformations.jl Spherical type. The UFF Point defines the azimuth θ as the angle from the point location to the YZ plane. The elevation ϕ is the angle from the point location to the XZ plane.

USTB.UFF.PointFromCartesian — Type

(::PointFromCartesian)(x::AbstractVector)

Transformation functor to map 3D Cartesian cordinates into UFFs Point type. The conversion for $[x, y, z]$ is given by

\[\begin{aligned} r &= \sqrt{x^2+y^2+z^2} \\ \theta &= \text{atan}(x, z) \\ \phi &= \text{asin}(y, r) \end{aligned}\]

USTB.UFF.CartesianFromPoint — Type

(::CartesianFromPoint)(x::Point)

Transformation functor to map UFF Point into Cartesian coordinates. The conversion for $[r, \theta, \phi]$ is given by

\[\begin{aligned} x &= r\cdot\sin(\theta)\cdot\cos(\phi) \\ y &= r\cdot\sin(\phi) \\ z &= r\cdot\cos(\theta)\cdot\cos(\phi) \end{aligned}\]

Wavefronts

USTB.UFF.Wavefront.WavefrontType — Type

WavefrontType(x::Integer)

Enumeration for wave types. Exported through USTB.UFF submodule. Available options are

Wavefront	Value

Plane	0

Spherical	1

Photoacustic	2

See also WAVE

TODO: Cross link WAVE

Windows

USTB.UFF.Window.WindowType — Type

WindowType(x::Integer)

Enumeration for window types. Available options are

Window Type	Value

None	0

Boxcar	1

Flat	1

Rectangular	1

Hanning	2

Hamming	3

Tukey25	4

Tukey50	5

Tukey75	6

Tukey80	7

Sta	7

Scanline	8

See also PULSE, BEAM, PHANTOM, PROBE

TODO: Link up PULSE

TODO: Link up BEAM

TODO: Link up PHANTOM

TODO: Link up PROBE